+++ src/Command/PrepareImages.fs
@@ -16,6 +16,9 @@ module PrepareCommand =
         Option.optional "exclude-list" (Some "x") "Text file contains a list of files you want to exclude from searching (<c:yellow>one file at line</c>)." None
         Option.noValue "force" (Some "f") "If set, target directory will NOT be excluded, and images may be overwritten."
         Option.noValue "dry-run" None "If set, target directory will NOT be touched in anyway and images will only be sent to stdout."
+        Option.noValue "year" None "If set, target directory will have a sub-directory with year of the image created date."
+        Option.noValue "month" None "If set, target directory will have a sub-directory with month of the image created date."
+        Option.optional "fallback" None "If set, it will be used as a sub-directory in target directory for all files, which don't have other specific sub-directory." None
     ]
 
     let execute ((input, output): IO) =
@@ -36,11 +39,8 @@ module PrepareCommand =
             output.Table ["Source"; "Target"; "Exclude"; "Exclude list from"]
                 [[source; [target]; exclude; excludeList |> Option.toList] |> List.map (sprintf "%A")]
 
-        output.Section (
-            sprintf "Prepare to %s from:\n - %s"
-                target
-                (source |> String.concat "\n - ")
-        )
+        output.Section <| sprintf "Prepare images to %s" target
+        output.Message <| sprintf "From:\n - %s" (source |> String.concat "\n - ")
 
         {
             Source = source
@@ -50,6 +50,16 @@ module PrepareCommand =
                 | Input.IsSetOption "force" _ -> Override
                 | Input.IsSetOption "dry-run" _ -> DryRun
                 | _ -> Exclude
+            TargetSubdir =
+                match input, input with
+                | Input.IsSetOption "year" _, Input.IsSetOption "month" _ -> ByYearAndMonth
+                | Input.IsSetOption "year" _, _ -> ByYear
+                | Input.IsSetOption "month" _, _ -> ByMonth
+                | _ -> Flat
+            TargetSubdirFallback =
+                match input with
+                | Input.HasOption "fallback" fallback -> fallback |> OptionValue.stringValue
+                | _ -> None
             Exclude =
                 match exclude with
                 | [] -> None