+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Allow multiple sources
 - Allow to exclude images from a file
 - Allow to dry-run the preparation
+- Allow to create a target subdirs by image created year/month